Buffer now supports dynamic loading of different meshes
Removed a temporary value in MeshHolder and moved it to MeshObject.
Added a new triangle with different uv values (defines color for now).
The defining of buffers in GenerateBuffers() is probably done. Meshes
must be defined before a call to this function and be loaded into a
std::vector to load the meshes. GenerateBuffers() must always be called
before the render loop.